# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.6.0
  description: QuantForge API — multi-market data platform with paper trading and institutional risk engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/risk/profiles:
    get:
      operationId: listRiskProfiles
      tags: [risk]
      summary: List risk profiles
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Risk profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskProfileListResponse"
    post:
      operationId: createRiskProfile
      tags: [risk]
      summary: Create a risk profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRiskProfileRequest"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskProfileResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /v1/risk/profiles/{id}:
    get:
      operationId: getRiskProfile
      tags: [risk]
      summary: Get a risk profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Risk profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskProfileResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateRiskProfile
      tags: [risk]
      summary: Update a risk profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRiskProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskProfileResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/risk/strategies/{name}:
    get:
      operationId: getStrategyRiskScore
      tags: [risk]
      summary: Get latest risk score for a specific strategy
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Strategy risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyRiskScoreResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/risk/strategies/score:
    post:
      operationId: scoreStrategy
      tags: [risk]
      summary: Manually trigger risk scoring for one or all strategies
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreStrategyRequest"
      responses:
        "201":
          description: Scoring complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyRiskScoreListResponse"
        "422":
          $ref: "#/components/responses/InsufficientData"
...
components:
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InsufficientData:
      description: Not enough data to perform computation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    RiskProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [conservative, balanced, aggressive, research, custom]
        maxPositionSizePct:
          type: string
        maxPortfolioExposurePct:
          type: string
        maxDailyLossPct:
          type: string
        maxDrawdownPct:
          type: string
        maxLeverage:
          type: string
        concentrationLimitPct:
          type: string
        maxWeeklyLossPct:
          type: string
        minStrategyConfidenceScore:
          type: string
        maxOpenPositions:
          type: integer
        riskToleranceScore:
          type: integer
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, profileType, maxPositionSizePct, maxPortfolioExposurePct, maxDailyLossPct, maxDrawdownPct, isDefault, isActive, createdAt, updatedAt]
    CreateRiskProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        profileType:
          type: string
          enum: [conservative, balanced, aggressive, research, custom]
          default: custom
        maxPositionSizePct:
          type: number
          minimum: 0
          maximum: 100
          default: 10
        maxPortfolioExposurePct:
          type: number
          minimum: 0
          maximum: 100
          default: 80
        maxDailyLossPct:
          type: number
          minimum: 0
          maximum: 100
          default: 3
        maxDrawdownPct:
          type: number
          minimum: 0
          maximum: 100
          default: 20
        maxLeverage:
          type: number
          minimum: 0
          maximum: 10
          default: 1
        concentrationLimitPct:
          type: number
          minimum: 0
          maximum: 100
          default: 25
        maxWeeklyLossPct:
          type: number
          minimum: 0
          maximum: 100
          default: 7
        minStrategyConfidenceScore:
          type: number
          minimum: 0
          maximum: 100
          default: 0
        maxOpenPositions:
          type: integer
          minimum: 1
          maximum: 500
          default: 20
        riskToleranceScore:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        isDefault:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
      required: [name]
    UpdateRiskProfileRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        maxPositionSizePct:
          type: number
        maxPortfolioExposurePct:
          type: number
        maxDailyLossPct:
          type: number
        maxDrawdownPct:
          type: number
        concentrationLimitPct:
          type: number
        maxWeeklyLossPct:
          type: number
        isDefault:
          type: boolean
        isActive:
          type: boolean
    RiskProfileResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/RiskProfile"
      required: [profile]
    RiskProfileListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RiskProfile"
        total:
          type: integer
      required: [data, total]
...
    StrategyRiskScore:
      type: object
      properties:
        id:
          type: string
          format: uuid
        strategyName:
          type: string
        winRateScore:
          type: string
        drawdownScore:
          type: string
        sharpeScore:
          type: string
        consistencyScore:
          type: string
        tradeFrequencyScore:
          type: string
        exposureScore:
          type: string
        overallRiskScore:
          type: string
        healthScore:
          type: string
        confidenceScore:
          type: string
        sampleSize:
          type: integer
        calculatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, strategyName, overallRiskScore, healthScore, confidenceScore, sampleSize, calculatedAt, createdAt]
    StrategyRiskScoreResponse:
      type: object
      properties:
        score:
          $ref: "#/components/schemas/StrategyRiskScore"
      required: [score]
    StrategyRiskScoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StrategyRiskScore"
        total:
          type: integer
      required: [data, total]
    ScoreStrategyRequest:
      type: object
      properties:
        strategyName:
          type: string
...